Changed age unit to months, merged protocol.js with app.js, fixed validity of form, implemented setProfile message type
(Not yet tested)
