+++ index.html
@@ -63,7 +63,7 @@
                 </div>
 
                 <div id="controls-profile-height-cm" class="inactive">
-                    <input id="controls-profile-height-cm-cm" name="height" type="number" placeholder="cm" min="0" step="0.01" required>
+                    <input id="controls-profile-height-cm-cm" name="height" type="number" placeholder="cm" min="0" step="0.01">
                 </div>
             </p>
             <p>
